flask_advertisements/pyproject.toml
[tool.poetry]
name = "flask-adv"
version = "0.1.0"
description = ""
authors = ["Lusi_shch"]
readme = "README.md"
packages = [{include = "flask_advertisements"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.3"
psycopg2 = "^2.9.6"
flask-restx = "^1.1.0"
swagger-ui-bundle = "^0.0.9"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
